Module itron_asp3::abi

C API

Note: doc(cfg(...)) is not used in this module as different kernels might provide different items with conflicting names. This module’s documentation might not be useful unless you ran cargo doc with an appropriate kernel selection.

This documentation has been built with the following Cargo features: asp3 , dcre , rstr_task , messagebuf , ovrhdr , subprio

Structs

T_CALM

TOPPERS/ASP3 T_CALM

T_CCYC

TOPPERS/ASP3 T_CCYC

T_CDTQ

T_CFLG

T_CISR

TOPPERS/ASP3 and SOLID/FMP3 T_CISR

T_CMPF

T_CMTX

T_CPDQ

T_CSEM

T_CTSK

TOPPERS/ASP3 dynamic creation extension T_CTSK

T_ENFY_DTQ

T_ENFY_VAR

T_NFYINFO

T_NFY_DTQ

T_NFY_FLG

T_NFY_HDR

T_NFY_IVAR

T_NFY_SEM

T_NFY_TSK

T_NFY_VAR

T_RALM

TOPPERS/ASP3 T_RALM

T_RCYC

TOPPERS/ASP3 T_RCYC

T_RDTQ

T_RFLG

T_RMBF

T_RMPF

T_RMTX

T_ROVR

T_RPDQ

T_RSEM

T_RTSK

TOPPERS/ASP3 T_RTSK

acvct

アクセス許可ベクタ

Constants

E_BOVR

バッファオーバーフロー

E_CLS

待ちオブジェクトの状態変化

E_COMM

通信エラー

E_CTX

コンテキストエラー

E_DLT

待ちオブジェクトの削除

E_ID

不正ID番号

E_ILUSE

サービスコール不正使用

E_MACV

メモリアクセス違反

E_NOEXS

オブジェクト未生成

E_NOID

ID番号不足

E_NOMEM

メモリ不足

E_NORES

資源不足

E_NOSPT

未サポート機能

E_OACV

オブジェクトアクセス違反

E_OBJ

オブジェクト状態エラー

E_OK

正常終了

E_PAR

パラメータエラー

E_QOVR

キューイングオーバーフロー

E_RASTER

タスクの終了要求

E_RLWAI

待ち状態の強制解除

E_RSATR

予約属性

E_RSFN

予約機能コード

E_SYS

システムエラー

E_TMOUT

ポーリング失敗またはタイムアウト

E_WBLK

ノンブロッキング受付け

FALSE

Represents a falsy value.

TACP_KERNEL

カーネルドメインだけにアクセスを許可

TACP_SHARED

すべてのドメインからアクセスを許可

TA_ACT

タスクを起動された状態で生成

TA_CEILING

優先度上限プロトコル

TA_CLR

イベントフラグのクリア指定

TA_NOACTQUE

起動要求をキューイングしない

TA_NULL

オブジェクト属性を指定しない

TA_RSTR

制約タスク

TA_STA

周期通知を動作状態で生成

TA_TPRI

待ち行列をタスクの優先度順にする

TA_WMUL

複数の待ちタスク

TENFY_ACTTSK

タスクの起動

TENFY_INCVAR

変数のインクリメント

TENFY_SETFLG

イベントフラグのセット

TENFY_SETVAR

変数の設定

TENFY_SIGSEM

セマフォの返却

TENFY_SNDDTQ

データキューへの送信

TENFY_WUPTSK

タスクの起床

TIPM_ENAALL

割込み優先度マスク全解除

TMAX_RELTIM

相対時間（RELTIM）に指定できる最大値［NGKI0551］

TMO_FEVR

永久待ち

TMO_NBLK

ノンブロッキング

TMO_POL

ポーリング

TNFY_ACTTSK

タスクの起動

TNFY_HANDLER

タイムイベントハンドラの呼出し

TNFY_INCVAR

変数のインクリメント

TNFY_SETFLG

イベントフラグのセット

TNFY_SETVAR

変数の設定

TNFY_SIGSEM

セマフォの資源の返却

TNFY_SNDDTQ

データキューへの送信

TNFY_WUPTSK

タスクの起床

TPRI_INI

タスクの起動時優先度

TPRI_SELF

自タスクのベース優先度

TRUE

Represents a truthy value.

TSK_NONE

該当するタスクがない

TSK_SELF

自タスク指定

TTS_DMT

休止状態

TTS_RDY

実行可能状態

TTS_RUN

実行状態

TTS_SUS

強制待ち状態

TTS_WAI

待ち状態

TTS_WAS

二重待ち状態

TTW_DLY

時間経過待ち

TTW_FLG

イベントフラグ待ち

TTW_MPF

固定長メモリブロックの獲得待ち

TTW_MTX

ミューテックスのロック待ち状態

TTW_RDTQ

データキューからの受信待ち

TTW_RMBF

メッセージバッファからの受信待ち

TTW_RPDQ

優先度データキューからの受信待ち

TTW_SDTQ

データキューへの送信待ち

TTW_SEM

セマフォの資源獲得待ち

TTW_SLP

起床待ち

TTW_SMBF

メッセージバッファへの送信待ち

TTW_SPDQ

優先度データキューへの送信待ち

TWF_ANDW

イベントフラグのAND待ち

TWF_ORW

イベントフラグのOR待ち

Functions

acre_alm⚠

acre_cyc⚠

acre_dtq⚠

acre_flg⚠

acre_isr⚠

acre_mpf⚠

acre_mtx⚠

acre_pdq⚠

acre_sem⚠

acre_tsk⚠

act_tsk⚠

adj_tim⚠

can_act⚠

can_wup⚠

chg_ipm⚠

chg_pri⚠

chg_spr⚠

clr_flg⚠

clr_int⚠

del_alm⚠

del_cyc⚠

del_dtq⚠

del_flg⚠

del_isr⚠

del_mpf⚠

del_mtx⚠

del_pdq⚠

del_sem⚠

del_tsk⚠

dis_dsp⚠

dis_int⚠

dis_ter⚠

dly_tsk⚠

ena_dsp⚠

ena_int⚠

ena_ter⚠

ext_ker⚠

ext_tsk⚠

fch_hrt⚠

fsnd_dtq⚠

get_inf⚠

get_ipm⚠

get_lod⚠

get_mpf⚠

get_nth⚠

get_pri⚠

get_tid⚠

get_tim⚠

get_tst⚠

ini_dtq⚠

ini_flg⚠

ini_mbf⚠

ini_mpf⚠

ini_mtx⚠

ini_pdq⚠

ini_sem⚠

loc_cpu⚠

loc_mtx⚠

pget_mpf⚠

ploc_mtx⚠

pol_flg⚠

pol_sem⚠

prb_int⚠

prcv_dtq⚠

prcv_mbf⚠

prcv_pdq⚠

psnd_dtq⚠

psnd_mbf⚠

psnd_pdq⚠

ras_int⚠

ras_ter⚠

rcv_dtq⚠

rcv_mbf⚠

rcv_pdq⚠

ref_alm⚠

ref_cyc⚠

ref_dtq⚠

ref_flg⚠

ref_mbf⚠

ref_mpf⚠

ref_mtx⚠

ref_ovr⚠

ref_pdq⚠

ref_sem⚠

ref_tsk⚠

rel_mpf⚠

rel_wai⚠

rot_rdq⚠

rsm_tsk⚠

set_flg⚠

set_tim⚠

sig_sem⚠

slp_tsk⚠

snd_dtq⚠

snd_mbf⚠

snd_pdq⚠

sns_ctx⚠

sns_dpn⚠

sns_dsp⚠

sns_ker⚠

sns_loc⚠

sns_ter⚠

sta_alm⚠

sta_cyc⚠

sta_ovr⚠

stp_alm⚠

stp_cyc⚠

stp_ovr⚠

sus_tsk⚠

ter_tsk⚠

tget_mpf⚠

tloc_mtx⚠

trcv_dtq⚠

trcv_mbf⚠

trcv_pdq⚠

tslp_tsk⚠

tsnd_dtq⚠

tsnd_mbf⚠

tsnd_pdq⚠

twai_flg⚠

twai_sem⚠

unl_cpu⚠

unl_mtx⚠

wai_flg⚠

wai_sem⚠

wup_tsk⚠

xsns_dpn⚠

Type Definitions

ACPTN

アクセス許可パターン

ATR

オブジェクトの属性

ER

エラーコード

ER_BOOL

エラーコードまたは真偽値

ER_ID

エラーコードまたはID番号

ER_UINT

エラーコードまたは符号無し整数

EXCNO

CPU例外ハンドラ番号

EXINF

拡張情報

FLGPTN

イベントフラグのビットパターン

FN

機能コード

FP

プログラムの起動番地

HRTCNT

高分解能タイマのカウント値

ID

オブジェクトのID番号

INHNO

割込みハンドラ番号

INTNO

割込み番号

ISR

MB_T

管理領域を確保するためのデータ型

MODE

サービスコールの動作モード

MPF_T

固定長メモリプール領域を確保するための型

NonNullID

Non-null version of ID

NonZeroER

PRCTIM

プロセッサ時間［NGKI0573］

PRI

優先度

RELTIM

相対時間

STAT

オブジェクトの状態

SYSTIM

システム時刻

TASK

TMEHDR

TMO

タイムアウト指定

bool_t

int_t

自然なサイズの符号付き整数

uint_t

自然なサイズの符号無し整数

Unions

T_ENFY

T_NFYINFO::enfy

T_NFY

T_NFYINFO::nfy