Struct itron_asp3::messagebuffer::MessageBufferRef

pub struct MessageBufferRef<'a> { /* fields omitted */ }

This is supported on crate feature unstable and (crate features asp3 and messagebuf, or crate features solid_asp3 and messagebuf, or crate feature none) only.

A borrowed reference to a message buffer.

Implementations

impl MessageBufferRef<'_>

Object ID conversion

pub const unsafe fn from_raw_nonnull(id: NonNullID) -> Self

Construct a MessageBufferRef from a raw object ID.

Safety

See Object ID Wrappers.

pub const fn as_raw(self) -> ID

Get the raw object ID.

pub const fn as_raw_nonnull(self) -> NonNullID

Get the raw object ID as abi::NonNullID.

impl MessageBufferRef<'_>

Management

pub fn info(self) -> Result<Info, Error<InfoError>>

ref_mbf: Get the message buffer’s general information.

impl MessageBufferRef<'_>

Message Buffer Operations

pub fn send(
self,
message: &impl Unknown + ?Sized
) -> Result<(), Error<SendError>>

snd_mbf: Send a message to the message buffer. Blocks the current task if the message buffer is full.

pub fn send_timeout(
self,
message: &impl Unknown + ?Sized,
tmo: Timeout
) -> Result<(), Error<SendTimeoutError>>

tsnd_mbf: Send a message to the message buffer. Blocks the current task with timeout if the message buffer is full.

pub fn try_send(
self,
message: &impl Unknown + ?Sized
) -> Result<(), Error<TrySendError>>

psnd_mbf: Send a message to the message buffer. Fails and returns an error if the message buffer is full.

pub unsafe fn recv(
self,
out: &mut impl Unknown + ?Sized
) -> Result<usize, Error<RecvError>>

rcv_mbf: Receive a message from the message buffer. Blocks the current task if the message buffer is empty.

On success, the received message will be written to *out, and the message size, measured in bytes, will be returned.

Rationale

This method is named recv instead of receive following the suit of std::sync::mpsc::Receiver::recv and std::net::UdpSocket::recv.

Safety

*out must be large enough to fit the message. After this method overwrites *out with the received message, the resultant content of *out must be valid for its type.

pub unsafe fn recv_timeout(
self,
out: &mut impl Unknown + ?Sized,
tmo: Timeout
) -> Result<usize, Error<RecvTimeoutError>>

trcv_mbf: Receive a message to the message buffer. Blocks the current task with timeout if the message buffer is empty.

Safety

See Self::recv.

pub unsafe fn try_recv(
self,
out: &mut impl Unknown + ?Sized
) -> Result<usize, Error<TryRecvError>>

prcv_mbf: Receive a message to the message buffer. Fails and returns an error if the message buffer is empty.

Safety

See Self::recv.

pub fn initialize(self) -> Result<(), Error<InitializeError>>

ini_mbf: Initialize the message buffer.